+++ tests/test_webapi.py
@@ -1,6 +1,7 @@
 # -*- encoding: utf-8 -*-
-
+import json
 from io import StringIO
+from uuid import UUID
 
 from django.urls import reverse
 from rest_framework import status
@@ -44,7 +45,6 @@ def create_format_spec(client, document_ref):
 
     response = client.post(
         url,
-        format="json",
         data={
             "document_ref": document_ref,
             "title": "My dummy document",
@@ -75,7 +75,6 @@ def create_quantity_spec(client, name, entity, format_spec, data_files=[]):
 
     response = client.post(
         url,
-        format="json",
         data={
             "name": name,
             "parent_entity": entity,
@@ -95,17 +94,19 @@ def create_data_file_spec(
 
     data_file = StringIO("1,2,3,4,5")
 
+    # Since we are sending "file_data", we cannot use
+    # format="json" here. Because of this, we need to
+    # manually convert `metadata` into a string, otherwise
+    # the `post` method would not be able to send both
+    # the metadata (a nested structure) and the file data.
     response = client.post(
         url,
-        format="json",
         data={
             "name": name,
-            "metadata": metadata,
+            "metadata": metadata if isinstance(metadata, str) else json.dumps(metadata),
             "quantity": quantity,
             "spec_version": spec_version,
             "release_tags": release_tags,
-        },
-        files={
             "file_data": data_file,
         },
     )
@@ -117,15 +118,22 @@ def create_release_spec(client, tag, comment="", data_files=[]):
 
     url = reverse("release-list")
 
+    from io import StringIO
+
+    release_document = StringIO("Contents of the release document")
+    release_document.name = "reldoc.txt"
+
     response = client.post(
         url,
-        format="json",
         data={
             "tag": tag,
             "comment": comment,
             "data_files": data_files,
+            "release_document_mime_type": "text/plain",
+            "release_document": release_document,
         },
     )
+
     return response
 
 
@@ -162,6 +170,18 @@ class EntityTests(APITestCase):
         self.assertEqual(Entity.objects.count(), 1)
         self.assertEqual(Entity.objects.get().name, "test_entity")
 
+        response = self.client.get("/tree/test_entity/")
+        self.assertEqual(response.status_code, status.HTTP_302_FOUND)
+
+        # Since we got HTTP 302, this is a redirect and we must follow the alias
+        response = self.client.get(response.url)
+        entity_json = response.json()
+        self.assertEqual(UUID(entity_json["uuid"]), Entity.objects.get().uuid)
+
+    def test_look_for_nonexistent_quantity(self):
+        response = self.client.get("/tree/this_entity_does_not_exist")
+        self.assertEqual(response.status_code, status.HTTP_301_MOVED_PERMANENTLY)
+
 
 class QuantityTests(APITestCase):
     def setUp(self):
@@ -188,6 +208,18 @@ class QuantityTests(APITestCase):
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
         self.assertEqual(Quantity.objects.count(), 1)
         self.assertEqual(Quantity.objects.get().name, "test_quantity")
+
+        response = self.client.get("/tree/test_entity/test_quantity/")
+        self.assertEqual(response.status_code, status.HTTP_302_FOUND)
+
+        # Since we got HTTP 302, this is a redirect and we must follow the alias
+        response = self.client.get(response.url)
+        quantity_json = response.json()
+        self.assertEqual(UUID(quantity_json["uuid"]), Quantity.objects.get().uuid)
+
+    def test_look_for_nonexistent_quantity(self):
+        response = self.client.get("/tree/test_entity/this_does_not_exist")
+        self.assertEqual(response.status_code, status.HTTP_301_MOVED_PERMANENTLY)
 
 
 class DataFileTests(APITestCase):
@@ -295,10 +327,14 @@ class ReleaseTests(APITestCase):
                 "data_files": [self.datafile_response.data["url"]],
             },
         )
+        assert response.status_code == status.HTTP_200_OK
+
         response = self.client.get(rel_url)
         json = response.json()
         self.assertEqual(json["tag"], "v1.0")
         self.assertEqual(len(json["data_files"]), 1)
+        self.assertEqual(json["release_document_mime_type"], "text/plain")
+        self.assertTrue(json["release_document"] is not None)
 
         # Try to access the data file through the release tag
         response = self.client.get(
@@ -306,6 +342,11 @@ class ReleaseTests(APITestCase):
         )
         self.assertEqual(response.status_code, status.HTTP_302_FOUND)
         assert response.url in self.datafile_response.data["url"]
+
+        # Download the release document
+
+        response = self.client.get("/browse/releases/v1.0/document/")
+        self.assertEqual(response.content, b"Contents of the release document")
 
 
 class AuthenticateTest(APITestCase):