+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # HEAD
 
+-   Add a new API to access entities and quantities through their path [#98](https://github.com/ziotom78/instrumentdb/pull/98)
+
+-   Make the download of specification documents easier [#97](https://github.com/ziotom78/instrumentdb/pull/97)
+
+-   Implement release documents, support Django 4 [#96](https://github.com/ziotom78/instrumentdb/pull/96)
+
+-   The URL `/api/releases` now uses pagination [#95](https://github.com/ziotom78/instrumentdb/pull/95)
+
 -   The tree of entities is now collapsible [#94](https://github.com/ziotom78/instrumentdb/pull/94)
 
 -   Extend test coverage, fix issues in the code [#93](https://github.com/ziotom78/instrumentdb/pull/93)